Fixed graphic for Metric Results not displaying correctly
The graphic for Metric results previously did not have a set size.
Changed it so that it fills the whole space available, and made it
responsive to window resizing. That broke some particular cases of
histories with few changes and number of points that wasn't very small.

In addition, made the graphic display a single point instead of an
unhelpful message.