Repository routes decoupled from Project
new and create action can be acessed both with or without a project. This
enables to have repositories the two ways.

The behaviour for the back had to change from going back to a specific project
page. To the all projects pages since some repositories may not have a
project_id.

With the routes, many helpers had to get fixed together with acceptance
tests steps (for this a new step was created in order to be able to create
RepositoryAttributes).

Finally the JS for repository state retrieval needed fixes since it builds the
route withot any helper. Same thing for the JS that prevents infinite state
queries.

Signed off by: Rafael Reggiani Manzo <rr.manzo@gmail.com>