Refactor MetricConfiguration controllers
Move as much code as possible to BaseMetricConfigurationsController
instead of repeating it ad-hoc in the concrete classes. Most of the
concrete code has gone away except for some small overrides:

- Hotspot's does not have a 'new' action, and therefore can't redirect
to it. It also does not accept any MetricConfiguration parameters.
- Metric's only accepts 'weight' and 'aggregation_form'
MetricConfiguration parameters. It also needs to fetch ReadingGroups to
display in the 'choose_metric' action.
- Compound's needs to force the Metric type when creating the metric,
and must accept parameters for Metric ('name', 'description', 'script',
'scope', 'code') in addition to MetricConfiguration's 'weight'.

Additionally:
- Update cache clearing to clear caches for Hotspot and Tree
configurations
- Verify the existence of the KalibroConfiguration and use the object
that was fetched whenever necessary, instead of only using the id.
- Ensure the KalibroConfiguration and ReadingGroup ids match between the
values inside and outside the 'metric_configuration' parameter.
Previously it was possible to send one in the URL and update another,
which might even have been a security issue.
- Allow sending either 'metric_code' or 'metric_name' as a parameter.
Both are used in different places in the code, but making them uniform
is a job for a later time.

Signed off by: Daniel Miranda <danielkza2@gmail.com>