Enable CSRF protection in notify_push acceptance tests
A bug in notify_push was missed due to CSRF protection being disabled in
the test environment. Fix that by adding a tag that can enabled it in
some features.

We might eventually want to enable it for all acceptance tests to make
sure other cases like this don't show up.