Refactor Webhook support to allow verification of parameters
Ensure webhook requests are only accepted after proper verification of
expected headers, repository address and branch, using the newly
implemented library.

To that effect, split up the logic to extract that information from
requests to it's own class, and use it in the notify_push action.

Also, add factories to make better testing of the notify_push action
possible, such as a factory similar to a real request from Gitlab, and
one for the Kalibro Client repository on Gitlab.

Adding other hook providers should be quite a lot easier after this is
applied.