Update header and request handling in tests
The repository/notify_push feature was making post request directly
using Rack::Test methods, which is not optimal when we have Capybara to
handle all the driver setup and flexibility for us.

Change it to use Capybara methods, and fix all the steps accordingly. To
make things easier, add some helper methods to deal with headers in
different drivers.

Also fix a small clarity issue in the notify_push unity tests that also
have to set headers.