Make travis fail if coverage is less the 100%
This is meant to ensure we remember when we commit changes without
tests. Also, to be able to run all the tests locally without worrying
about that, we make the minimum coverage only required if the
environment variable 'CI' is set to true. In practice, this means we
only check minimun coverage if we are building the environment on
travis.

Also remove some default configurations for Rails on Simplecov.

Updates CHANGELOG